Subjects ALL

+++ src/main/java/com/peter/kist/repository/SubjectRepository.java
@@ -3,5 +3,5 @@ package com.peter.kist.repository;
 import com.peter.kist.model.Subject;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubjectRepository extends JpaRepository<Subject, Long> {
+public interface SubjectRepository extends JpaRepository<Subject, Integer> {
 }
